Fix: create room dialog Fix: create room dialog animation

+++ app/src/main/java/com/iriras/aplikasichatfirebase/util/helper/FxHelper.kt
@@ -6,10 +6,10 @@ import android.content.Context
  * Created by irfan on 8/24/17.
  */
 
-fun Int.pxToDp(context: Context, source: Int): Int {
-    return (source / context.resources.displayMetrics.density).toInt()
+fun Float.pxToDp(context: Context, source: Float): Float {
+    return source / context.resources.displayMetrics.density
 }
 
-fun Int.dpToPx(context: Context, source: Int): Int {
-    return (source * context.resources.displayMetrics.density).toInt()
+fun Float.dpToPx(context: Context, source: Float): Float {
+    return source * context.resources.displayMetrics.density
 }
